% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_custom_render.R
\name{custom_render}
\alias{custom_render}
\title{Custom report rendering}
\usage{
custom_render(template = "?", data = NULL, ...)
}
\arguments{
\item{template}{a template. Use "?" to list available template.}

\item{data}{data passed to [whisker::whisker.render()].}

\item{...}{further arguments passed to [rmarkdown::render()].}
}
\value{
The path of the output files or a vector of template names,
invisibly.
}
\description{
A function to render R Makdown templates found in `inst/
templates`.
}
\examples{
\dontrun{
custom_render("report1",
    data = list(title = "templex"),
    envir = list(geom = 1)
)
}
}
