% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mb_tracks.R
\name{get_mb_tracks}
\alias{get_mb_tracks}
\title{Retrieve movebank tracks}
\usage{
get_mb_tracks(
  mb_user = Sys.getenv("MOVEBANK_USER"),
  mb_password = Sys.getenv("MOVEBANK_PW"),
  study_id = NULL
)
}
\arguments{
\item{mb_user}{movebank username.}

\item{mb_password}{movebank password.}

\item{study_id}{movebank studyId.}
}
\value{
Return a data.frame of 7 columns:
- tag_id: Individual tag id
- band_id: Individual band id
- lon: Longitude of the location
- lat: Latitude of the location
- datetime: Datetime of the location
- species: Species latin name
- vernacular: Species english name
}
\description{
Retrieve tracks from movebank project Id.
}
