% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nunatsiavut_gulls}
\alias{nunatsiavut_gulls}
\title{Year-round Tracking of Breeding Gulls in Nunatsiavut}
\format{
A data frame with 40,477 rows and 40 columns
}
\usage{
nunatsiavut_gulls
}
\description{
This dataset represents all locations of the movebank project ID 2854587542
This static extraction has been made February, 15th, 2024
}
\keyword{datasets}
