% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_gs_parquet.R
\name{sync_gs_parquet}
\alias{sync_gs_parquet}
\title{Write parquet tracks in GCS bucket}
\usage{
sync_gs_parquet(
  bucket = NULL,
  auth_gcs_file_path = "./nunatsiavut-birds-f33436183827.json",
  ...
)
}
\arguments{
\item{bucket}{GCS bucket name.}

\item{auth_gcs_file_path}{movebank password.}

\item{...}{arguments passed to get_mb_tracks function}
}
\description{
Write parquet files with animals tracks into the target GCS bucket.
}
